Open api(Swagger added), dto work was implemented, mapstruct added

+++ backend/src/main/java/com/relucky/backend/models/Employee.java
@@ -21,5 +21,5 @@ public class Employee {
     private String surname;
     private Integer age;
     private String email;
-    private Double salary;
+    private Integer salary;
 }
